pyproject.toml
[project]
name = "ekomurojat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-unfold>=0.74.1",
    "environs>=14.5.0",
    "gdal==3.8.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
